+++ .husky/pre-commit
@@ -4,5 +4,5 @@
 # npm run test
 npm run lint:fix
 npm run format
-npm run build
+# npm run build
 git add .
